requirements-test.txt
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
requests-mock==1.11.0
